rename: Move package directory from 'source' to 'rich_chat'
- Rename package directory from 'source' to 'rich_chat' for proper naming conventions
- Improve organization and separation of concerns in the codebase
- Create a separate 'chat_history.py' module for the `ChatHistory` class
- Move helper function 'keybinding' from 'rich-chat.py' to 'chat_history.py'
- Move `ChatHistory` class from 'rich-chat.py' to 'chat_history.py' and remove unused imports in 'rich-chat.py'
- Add import statement for 'ChatHistory' in 'rich-chat.py'
- Rename package directory for proper naming conventions
- Improve organization and separation of concerns
- Consolidate related functionality into separate modules
